test(skills): extract skill discovery tests into an owned test seam
Move the 1,470-line inline tests module out of skills/mod.rs into
skills/tests.rs (mechanical motion; tests reference parent items
through super:: paths so the move is verbatim apart from one dedent
level). The production owner shrinks to 1,346 lines and the counted
aggregate drops by the same amount, since tests.rs paths are excluded
from the source-structure budget.

Verified: skills suites 195/195 PASS from the new location; fmt clean.
